Remove Vue SFC dependency on plugin
