+++ package.json
@@ -7,7 +7,6 @@
     },
     "devDependencies": {
         "@tailwindcss/forms": "^0.5.2",
-        "@vitejs/plugin-vue": "^5.1.2",
         "@types/toastify-js": "^1.12.3",
         "alpinejs": "^3.4.2",
         "autoprefixer": "^10.4.2",
