+++ resources/css/app.css
@@ -54,3 +54,13 @@
         color: var(--brand-secondary);
     }
 }
+
+.fade-enter-active,
+.fade-leave-active {
+    transition: opacity 0.2s ease;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+    opacity: 0;
+}
