+++ vite.config.js
@@ -1,6 +1,5 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
-import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
     server: {
@@ -17,8 +16,12 @@ export default defineConfig({
             usePolling: true,
         },
     },
+    resolve: {
+        alias: {
+            vue: 'vue/dist/vue.esm-bundler.js',
+        },
+    },
     plugins: [
-        vue(),
         laravel({
             input: [
                 'resources/js/app.js',
